% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baidu.R
\name{en2cn}
\alias{en2cn}
\alias{cn2en}
\alias{baidu_translate}
\title{baidu_translate}
\usage{
en2cn(x)

cn2en(x)

baidu_translate(x, from = "en", to = "zh")
}
\arguments{
\item{x}{query sentence}

\item{from}{source language, i.e., the language to be translated}

\item{to}{target language, i.e., the language to be translated into}
}
\value{
the translated sentence
}
\description{
Translate query sentence
}
\details{
This function use the Baidu fanyi API to translate query sentences
}
\author{
Guangchuang Yu
}
