% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanyi-package.R
\docType{package}
\name{fanyi-package}
\alias{fanyi}
\alias{fanyi-package}
\title{fanyi: Translate Words or Sentences via Online Translators}
\description{
Useful functions to translate text for multiple languages using online translators. For example, by translating error messages and descriptive analysis results into a language familiar to the user, it enables a better understanding of the information, thereby reducing the barriers caused by language. It offers several helper functions to query gene information to help interpretation of interested genes (e.g., marker genes, differential expression genes), and provides utilities to translate 'ggplot' graphics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuLab-SMU/fanyi (devel)}
  \item \url{https://www.biorxiv.org/content/10.1101/2023.12.21.572729 (manuscript)}
  \item Report bugs at \url{https://github.com/YuLab-SMU/fanyi/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Guannan Chen \email{gnchen@hrbmu.edu.cn} [contributor]
}

}
\keyword{internal}
