% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.r
\name{fanyi_has_appkey}
\alias{fanyi_has_appkey}
\title{fanyi_has_appkey}
\usage{
fanyi_has_appkey(source)
}
\arguments{
\item{source}{one of the supported translate services}
}
\value{
logical value
}
\description{
test whether there has an appkey setting for the specific source
}
