% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-summary.r
\name{gene_summary}
\alias{gene_summary}
\title{gene_summary}
\usage{
gene_summary(entrez)
}
\arguments{
\item{entrez}{entrez gene IDs}
}
\value{
A data frame containing the information
}
\description{
query gene information from NCBI
}
\details{
This function query gene information (including gene name, description and summary) from NCBI Gene database
}
\examples{
\dontrun{
gene_summary('1236')
}
}
\author{
Guangchuang Yu
}
