% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{get_translate_text}
\alias{get_translate_text}
\title{get_translate_text}
\usage{
get_translate_text(response)
}
\arguments{
\item{response}{return from the online translator}
}
\value{
translated text
}
\description{
access translated text from online translator response
}
