% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name2sci.r
\name{name2sci}
\alias{name2sci}
\title{name2sci}
\usage{
name2sci(x)
}
\arguments{
\item{x}{common names}
}
\value{
corresponding scientific names
}
\description{
convert species common name to scientific name
}
\details{
The function query species common name via the NCBI Taxonomy database and return corresponding scientific name
}
\examples{
\dontrun{
name2sci("tiger")
}
}
