% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-summary.r
\name{symbol2entrez}
\alias{symbol2entrez}
\alias{search_gene}
\title{search_gene}
\usage{
symbol2entrez(x, organism = "Homo sapiens")

search_gene(x, organism = "Homo sapiens")
}
\arguments{
\item{x}{terms to search}

\item{organism}{correpsonding organism of the input terms}
}
\value{
A data frame with TERM and ENTREZ columns
}
\description{
search genes and return corresponding entrez ids
}
\details{
This function query genes (e.g., symbols) from NCBI Gene database and return corresponding entrez gene IDs
}
\author{
Guangchuang Yu
}
