% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baidu.R
\name{set_translate_option}
\alias{set_translate_option}
\title{set_translate_option}
\usage{
set_translate_option(appid, key, source = "baidu")
}
\arguments{
\item{appid}{appid}

\item{key}{app key}

\item{source}{translation engine, currently only 'baidu' is supported}
}
\description{
set appid and key of translation engine
}
\details{
This function allows users to use their own appid and key
}
\author{
Guangchuang Yu
}
