% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.r
\name{set_translate_source}
\alias{set_translate_source}
\title{set source of online translator service}
\usage{
set_translate_source(source)
}
\arguments{
\item{source}{translation engine}
}
\value{
No return value, called for side effects
}
\description{
This function allows users to set the default source for \code{translate()} function
}
\examples{
set_translate_source("baidu")
}
\author{
Guangchuang Yu
}
