% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-summary.r
\name{symbol2entrez}
\alias{symbol2entrez}
\title{symbol2entrez}
\usage{
symbol2entrez(symbols, organism = "Homo sapiens")
}
\arguments{
\item{symbols}{gene symbols}

\item{organism}{correpsonding organism of the gene symbols}
}
\value{
A data frame with SYMBOL and ENTREZ columns
}
\description{
convert gene symbol to entrez id
}
\details{
This function query gene symbols from NCBI Gene database and return corresponding entrez gene IDs
}
\author{
Guangchuang Yu
}
