% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.r
\name{translate_ggplot}
\alias{translate_ggplot}
\title{translate_ggplot}
\usage{
translate_ggplot(plot, axis = "xy", from = "en", to = "zh")
}
\arguments{
\item{plot}{a ggplot object to be translated}

\item{axis}{one of 'x', 'y' or 'xy' to select axis labels to be translated}

\item{from}{source language, i.e., the language to be translated}

\item{to}{target language, i.e., the language to be translated into}
}
\value{
a translated ggplot object
}
\description{
Translate axis label of a ggplot
}
\details{
This function use the \code{translate()} function to translate axis labels of a ggplot
}
\author{
Guangchuang Yu
}
