% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.r
\name{ydict}
\alias{ydict}
\title{ydict}
\usage{
ydict(word, web = FALSE)
}
\arguments{
\item{word}{word to check}

\item{web}{whether open youdao dict in web browser}
}
\value{
interpretation from youdao dictionary
}
\description{
check English word in Youdao dictionary
}
\examples{
\dontrun{
ydict('panda') 
ydict(tiger) # unquoted word is supported

# if using a word stored in a variable
#
x <- 'panda'
ydict(!!rlang::sym(x))

}
}
