% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.r
\name{set_translate_option}
\alias{set_translate_option}
\title{set_translate_option}
\usage{
set_translate_option(
  appid,
  key,
  source = "baidu",
  region = "southeastasia",
  user_dict = NULL,
  user_model = "turbo"
)
}
\arguments{
\item{appid}{appid, "bing translate" will not use this input.}

\item{key}{app key}

\item{source}{translation engine}

\item{region}{this is for bing use only, translation engine location, depends on your Azure service setting}

\item{user_dict}{user defined dictionary ID, only used for 'source = "youdao"'}

\item{user_model}{selected model used in for 'source = "chatglm"', one of \code{turbo}, \code{pro}, \code{standard} and \code{lite}}
}
\value{
No return value, called for side effects
}
\description{
set appid and key of translation engine
}
\details{
This function allows users to use their own appid and key
}
\author{
Guangchuang Yu
}
