% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baidu.R, R/translate.r
\name{baidu_translate}
\alias{baidu_translate}
\alias{en2cn}
\alias{cn2en}
\alias{translate}
\title{translate}
\usage{
baidu_translate(x, from = "en", to = "zh")

en2cn(x)

cn2en(x)

translate(x, from = "en", to = "zh")
}
\arguments{
\item{x}{query sentences}

\item{from}{source language, i.e., the language to be translated}

\item{to}{target language, i.e., the language to be translated into}
}
\value{
the translated sentences
}
\description{
Translate query sentences
}
\details{
This function use online translator API (e.g., Baidu fanyi) to translate query sentences
}
\examples{
library(fanyi)
## set your appid and key once in your R session
#
# set_translate_option(appid = 'your_appid', key = 'your_key')
#
# translate('hello world', from = 'en', to = 'zh')
}
\author{
Guangchuang Yu
}
